requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
pydantic
